Move UnexpectedResponseError to models
